requirements-dev.txt
-i https://pypi.org/simple
anyio==4.11.0; python_version >= '3.9'
appnope==0.1.4; python_version >= '3.6'
argon2-cffi==25.1.0; python_version >= '3.8'
argon2-cffi-bindings==25.1.0; python_version >= '3.9'
arrow==1.4.0; python_version >= '3.8'
asttokens==3.0.0; python_version >= '3.8'
async-lru==2.0.5; python_version >= '3.9'
attrs==25.4.0; python_version >= '3.9'
babel==2.17.0; python_version >= '3.8'
beautifulsoup4==4.14.2; python_full_version >= '3.7.0'
bleach[css]==6.2.0; python_version >= '3.9'
certifi==2025.10.5; python_version >= '3.7'
cffi==2.0.0; python_version >= '3.9'
cfgv==3.4.0; python_version >= '3.8'
charset-normalizer==3.4.4; python_version >= '3.7'
comm==0.2.3; python_version >= '3.8'
coverage==7.11.0; python_version >= '3.10'
debugpy==1.8.17; python_version >= '3.8'
decorator==5.2.1; python_version >= '3.8'
defusedxml==0.7.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
distlib==0.4.0
executing==2.2.1; python_version >= '3.8'
fastjsonschema==2.21.2
filelock==3.20.0; python_version >= '3.10'
fqdn==1.5.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
h11==0.16.0; python_version >= '3.8'
httpcore==1.0.9; python_version >= '3.8'
httpx==0.28.1; python_version >= '3.8'
identify==2.6.15; python_version >= '3.9'
idna==3.11; python_version >= '3.8'
ipykernel==7.0.1; python_version >= '3.10'
ipython==9.6.0; python_version >= '3.11'
ipython-pygments-lexers==1.1.1; python_version >= '3.8'
ipywidgets==8.1.7; python_version >= '3.7'
isoduration==20.11.0; python_version >= '3.7'
jedi==0.19.2; python_version >= '3.6'
jinja2==3.1.6; python_version >= '3.7'
json5==0.12.1; python_full_version >= '3.8.0'
jsonpointer==3.0.0; python_version >= '3.7'
jsonschema[format-nongpl]==4.25.1; python_version >= '3.9'
jsonschema-specifications==2025.9.1; python_version >= '3.9'
jupyter==1.1.1
jupyter-client==8.6.3; python_version >= '3.8'
jupyter-console==6.6.3; python_version >= '3.7'
jupyter-core==5.9.1; python_version >= '3.10'
jupyter-events==0.12.0; python_version >= '3.9'
jupyter-lsp==2.3.0; python_version >= '3.8'
jupyter-server==2.17.0; python_version >= '3.9'
jupyter-server-terminals==0.5.3; python_version >= '3.8'
jupyterlab==4.4.9; python_version >= '3.9'
jupyterlab-pygments==0.3.0; python_version >= '3.8'
jupyterlab-server==2.27.3; python_version >= '3.8'
jupyterlab-widgets==3.0.15; python_version >= '3.7'
lark==1.3.0; python_version >= '3.8'
markupsafe==3.0.3; python_version >= '3.9'
matplotlib-inline==0.1.7; python_version >= '3.8'
mistune==3.1.4; python_version >= '3.8'
nbclient==0.10.2; python_full_version >= '3.9.0'
nbconvert==7.16.6; python_version >= '3.8'
nbformat==5.10.4; python_version >= '3.8'
nest-asyncio==1.6.0; python_version >= '3.5'
nodeenv==1.9.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
notebook==7.4.7; python_version >= '3.8'
notebook-shim==0.2.4; python_version >= '3.7'
packaging==25.0; python_version >= '3.8'
pandocfilters==1.5.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
parso==0.8.5; python_version >= '3.6'
pexpect==4.9.0; sys_platform != 'win32' and sys_platform != 'emscripten'
platformdirs==4.5.0; python_version >= '3.10'
pre-commit==4.3.0; python_version >= '3.9'
prometheus-client==0.23.1; python_version >= '3.9'
prompt-toolkit==3.0.52; python_version >= '3.8'
psutil==7.1.1; python_version >= '3.6'
ptyprocess==0.7.0; os_name != 'nt'
pure-eval==0.2.3
pycparser==2.23; python_version >= '3.8'
pygments==2.19.2; python_version >= '3.8'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
python-json-logger==4.0.0; python_version >= '3.8'
pyyaml==6.0.3; python_version >= '3.8'
pyzmq==27.1.0; python_version >= '3.8'
referencing==0.37.0; python_version >= '3.10'
requests==2.32.5; python_version >= '3.9'
rfc3339-validator==0.1.4; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
rfc3986-validator==0.1.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
rfc3987-syntax==1.1.0; python_version >= '3.9'
rpds-py==0.27.1; python_version >= '3.9'
ruff==0.14.1; python_version >= '3.7'
send2trash==1.8.3; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
setuptools==80.9.0; python_version >= '3.9'
six==1.17.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
sniffio==1.3.1; python_version >= '3.7'
soupsieve==2.8; python_version >= '3.9'
stack-data==0.6.3
terminado==0.18.1; python_version >= '3.8'
tinycss2==1.4.0; python_version >= '3.8'
tornado==6.5.2; python_version >= '3.9'
traitlets==5.14.3; python_version >= '3.8'
typing-extensions==4.15.0; python_version >= '3.9'
tzdata==2025.2; python_version >= '2'
uri-template==1.3.0; python_version >= '3.7'
urllib3==2.5.0; python_version >= '3.9'
virtualenv==20.35.3; python_version >= '3.8'
wcwidth==0.2.14; python_version >= '3.6'
webcolors==24.11.1; python_version >= '3.9'
webencodings==0.5.1
websocket-client==1.9.0; python_version >= '3.9'
widgetsnbextension==4.0.14; python_version >= '3.7'
aiohappyeyeballs==2.6.1; python_version >= '3.9'
aiohttp==3.13.1; python_version >= '3.9'
aiosignal==1.4.0; python_version >= '3.9'
annotated-types==0.7.0; python_version >= '3.8'
azure-core==1.36.0; python_version >= '3.9'
azure-storage-blob==12.27.0; python_version >= '3.9'
click==8.3.0; python_version >= '3.10'
cryptography==46.0.3; python_version >= '3.8' and python_full_version not in '3.9.0, 3.9.1'
datasets==3.6.0; python_full_version >= '3.9.0'
dill==0.3.8; python_version >= '3.8'
frozenlist==1.8.0; python_version >= '3.9'
fsspec[http]==2025.3.0; python_version >= '3.8'
gitdb==4.0.12; python_version >= '3.7'
gitpython==3.1.45; python_version >= '3.7'
hf-xet==1.1.10; python_version >= '3.8'
huggingface-hub==0.35.3; python_full_version >= '3.8.0'
iniconfig==2.3.0; python_version >= '3.10'
isodate==0.7.2; python_version >= '3.7'
mpmath==1.3.0
multidict==6.7.0; python_version >= '3.9'
multiprocess==0.70.16; python_version >= '3.8'
networkx==3.5; python_version >= '3.11'
numpy==2.3.4; python_version >= '3.11'
pandas==2.3.3; python_version >= '3.9'
pluggy==1.6.0; python_version >= '3.9'
propcache==0.4.1; python_version >= '3.9'
protobuf==6.33.0; python_version >= '3.9'
pyarrow==21.0.0; python_version >= '3.9'
pydantic==2.12.3; python_version >= '3.9'
pydantic-core==2.41.4; python_version >= '3.9'
pytest==8.4.2; python_version >= '3.9'
pytest-mock==3.15.1; python_version >= '3.9'
pytz==2025.2
sentry-sdk==2.42.1; python_version >= '3.6'
smmap==5.0.2; python_version >= '3.7'
sympy==1.14.0; python_version >= '3.9'
tokenizers==0.22.1; python_version >= '3.9'
torch==2.9.0; python_version >= '3.10'
tqdm==4.67.1; python_version >= '3.7'
typing-inspection==0.4.2; python_version >= '3.9'
wandb==0.22.2; python_version >= '3.8'
xxhash==3.6.0; python_version >= '3.7'
yarl==1.22.0; python_version >= '3.9'
